Change task manager pool to be single threaded, there's no need to have it being a cached pool without any limit, specially because it currently is only being used for .ping().
Also change the QueueServerImpl, that's calling a .get all the time and waiting the response, when you could just simply process it or properly handle the exception when the ping finishes and returns a result.

Any ideas on this? I think this would be a good improvement, one thing is clear, in the old scenario that cached thread pool would cause lots of threads created on every ping, so the user would have to be all the time changing the delays that the servers get pinged in order to not have a high cpu.